requirements.txt
# Core framework
streamlit==1.54.0

# AI / LLM
google-generativeai==0.8.6

# Document parsing
docling==2.78.0
docling-core==2.69.0
docling-ibm-models==3.12.0
docling-parse==5.5.0

# Data handling
pandas==2.2.3
openpyxl==3.1.5

# Schema / validation
pydantic==2.12.5
pydantic-settings==2.13.1

# Environment
python-dotenv==1.1.1

# Visualization
plotly==6.5.2

# Image processing
pillow==11.1.0
